Add iSpindel calibration point management
