+++ gravity/views.py
@@ -7,7 +7,6 @@ from constance import config
 from django.contrib.auth.decorators import login_required
 from django.views.decorators.csrf import csrf_exempt
 from django.utils import timezone
-import json, socket, decimal
 from django.http import JsonResponse
 
 from app.models import BrewPiDevice
@@ -459,179 +458,7 @@ def gravity_manage(request, sensor_id):
 
         calibration_points = IspindelGravityCalibrationPoint.objects.filter(sensor=sensor)
         context['ispindel_calibration_points'] = calibration_points
-        ispindel_calibration_form = forms.IspindelGravityCalibrationPoint()
+        ispindel_calibration_form = forms.IspindelCalibrationPointForm(initial={'sensor': sensor})
         context['ispindel_calibration_form'] = ispindel_calibration_form
 
     return render(request, template_name='gravity/gravity_manage.html', context=context)
-
-
-@login_required
-@site_is_configured
-def gravity_ispindel_setup(request, sensor_id):
-    # TODO - Add user permissioning
-    # if not request.user.has_perm('app.edit_device'):
-    #     messages.error(request, 'Your account is not permissioned to edit devices. Please contact an admin')
-    #     return redirect("/")
-
-    try:
-        sensor = GravitySensor.objects.get(id=sensor_id)
-    except:
-        messages.error(request, u'Unable to load sensor with ID {}'.format(sensor_id))
-        return redirect('gravity_log_list')
-
-    if sensor.sensor_type != "ispindel":
-        messages.error(request, u"Sensor {} is not an iSpindel sensor".format(sensor.name))
-        return redirect('gravity_log_list')
-
-    # When we're configuring the iSpindel sensor, we need to tell it how to connect back to Fermentrack. Get the
-    # hostname, port, and IP address here so we can provide it to the user to enter.
-    fermentrack_host = request.META['HTTP_HOST']
-    try:
-        ais = socket.getaddrinfo(fermentrack_host, 0, 0, 0, 0)
-        ip_list = [result[-1][0] for result in ais]
-        ip_list = list(set(ip_list))
-        resolved_address = ip_list[0]
-    except:
-        # For some reason we failed to resolve the IP address of this host. Let's let the user know they'll have to
-        # figure it out.
-        resolved_address = "<The IP address of Fermentrack> (Unable to autodetect. Sorry.)"
-
-    return render(request, template_name='gravity/gravity_ispindel_setup.html',
-                  context={'active_device': sensor, 'fermentrack_host': fermentrack_host,
-                           'fermentrack_ip': resolved_address})
-
-
-@csrf_exempt
-def ispindel_handler(request):
-    if request.body is None:
-        # TODO - Log this
-        return JsonResponse({'status': 'failed', 'message': "No data in request body"}, safe=False,
-                            json_dumps_params={'indent': 4})
-
-    # import pprint
-    # with open('ispindel_json_output.txt', 'w') as logFile:
-    #     pprint.pprint(ispindel_data, logFile)
-
-    # As of the iSpindel firmware version 5.6.1, the json posted contains the following fields:
-    # {u'ID': 3003098,
-    #  u'angle': 77.4576,
-    #  u'battery': 4.171011,
-    #  u'gravity': 27.22998,
-    #  u'name': u'iSpindel123',
-    #  u'temperature': 24.75,
-    #  u'token': u'tokengoeshere'}
-
-    ispindel_data = json.loads(request.body)
-
-    try:
-        sensor = IspindelConfiguration.objects.get(name_on_device=ispindel_data['name'])
-    except:
-        # TODO - Log This
-        messages.error(request, u'Unable to load sensor with name {}'.format(ispindel_data['name']))
-        return JsonResponse({'status': 'failed', 'message': "Unable to load sensor with that name"}, safe=False,
-                            json_dumps_params={'indent': 4})
-
-
-    # Let's calculate the gravity using the coefficients stored in the ispindel configuration. This will allow us to
-    # reconfigure on the fly.
-    calculated_gravity = sensor.third_degree_coefficient * decimal.Decimal(ispindel_data['angle']**3)
-    calculated_gravity += sensor.second_degree_coefficient * decimal.Decimal(ispindel_data['angle']**2)
-    calculated_gravity += sensor.first_degree_coefficient * decimal.Decimal(ispindel_data['angle'])
-    calculated_gravity += sensor.constant_term
-
-    new_point = GravityLogPoint(
-        gravity=calculated_gravity,         # We're using the gravity we calc within Fermentrack
-        temp=ispindel_data['temperature'],
-        temp_format='C',                    # iSpindel devices always report temp in celsius
-        temp_is_estimate=False,
-        associated_device=sensor.sensor,
-        gravity_latest=calculated_gravity,
-        temp_latest=ispindel_data['temperature'],
-        extra_data=ispindel_data['angle'],
-    )
-
-    if sensor.sensor.active_log is not None:
-        new_point.associated_log = sensor.sensor.active_log
-
-    new_point.save()
-
-    # Set & save the 'extra' data points to redis (so we can load & use later)
-    if 'angle' in ispindel_data:
-        sensor.angle = ispindel_data['angle']
-    if 'ID' in ispindel_data:
-        sensor.ispindel_id = ispindel_data['ID']
-    if 'battery' in ispindel_data:
-        sensor.battery = ispindel_data['battery']
-    if 'gravity' in ispindel_data:
-        sensor.ispindel_gravity = ispindel_data['gravity']
-    if 'token' in ispindel_data:
-        sensor.token = ispindel_data['token']
-    sensor.save_extras_to_redis()
-
-    return JsonResponse({'status': 'ok', 'gravity': calculated_gravity}, safe=False, json_dumps_params={'indent': 4})
-
-
-@login_required
-@site_is_configured
-def gravity_ispindel_coefficients(request, sensor_id):
-    # TODO - Add user permissioning
-    # if not request.user.has_perm('app.edit_device'):
-    #     messages.error(request, 'Your account is not permissioned to edit devices. Please contact an admin')
-    #     return redirect("/")
-
-    try:
-        sensor = GravitySensor.objects.get(id=sensor_id)
-    except:
-        messages.error(request, u'Unable to load sensor with ID {}'.format(sensor_id))
-        return redirect('gravity_log_list')
-
-    if request.POST:
-        ispindel_coefficient_form = forms.IspindelCoefficientForm(request.POST)
-        if ispindel_coefficient_form.is_valid():
-            sensor.ispindel_configuration.third_degree_coefficient = ispindel_coefficient_form.cleaned_data['a']
-            sensor.ispindel_configuration.second_degree_coefficient = ispindel_coefficient_form.cleaned_data['b']
-            sensor.ispindel_configuration.first_degree_coefficient = ispindel_coefficient_form.cleaned_data['c']
-            sensor.ispindel_configuration.constant_term = ispindel_coefficient_form.cleaned_data['d']
-
-            sensor.ispindel_configuration.save()
-            messages.success(request, u"Coefficients updated")
-
-        else:
-            messages.error(request, u"Invalid coefficients provided")
-    else:
-        messages.error(request, u"No coefficients provided")
-
-    return redirect("gravity_manage", sensor_id=sensor_id)
-
-
-@login_required
-@site_is_configured
-def gravity_ispindel_calibration(request, sensor_id):
-    # TODO - Add user permissioning
-    # if not request.user.has_perm('app.edit_device'):
-    #     messages.error(request, 'Your account is not permissioned to edit devices. Please contact an admin')
-    #     return redirect("/")
-
-    try:
-        sensor = GravitySensor.objects.get(id=sensor_id)
-    except:
-        messages.error(request, u'Unable to load sensor with ID {}'.format(sensor_id))
-        return redirect('gravity_log_list')
-
-    if request.POST:
-        ispindel_coefficient_form = forms.IspindelCoefficientForm(request.POST)
-        if ispindel_coefficient_form.is_valid():
-            sensor.ispindel_configuration.third_degree_coefficient = ispindel_coefficient_form.cleaned_data['a']
-            sensor.ispindel_configuration.second_degree_coefficient = ispindel_coefficient_form.cleaned_data['b']
-            sensor.ispindel_configuration.first_degree_coefficient = ispindel_coefficient_form.cleaned_data['c']
-            sensor.ispindel_configuration.constant_term = ispindel_coefficient_form.cleaned_data['d']
-
-            sensor.ispindel_configuration.save()
-            messages.success(request, u"Coefficients updated")
-
-        else:
-            messages.error(request, u"Invalid coefficients provided")
-    else:
-        messages.error(request, u"No coefficients provided")
-
-    return redirect("gravity_manage", sensor_id=sensor_id)
